Add unsigned-integer overflow checking to UBSAN
Background:

We recently found a very subtle bug caused by unsigned integer overflow
in tablet_peer_mm_ops.cc (KUDU-2017). This patch was able to repro that
overflow from at least two different unit tests, so the value is high.
Additionally, it lets us be a little more aggressive about using
unsigned integers. I've never been much of a fan of the style guide's
ban.

A bunch of the overflow instances are expected, like hash functions and
random number generators. A few bugs were exposed, including a missing
statement in Schema::swap, and an accounting error in the thread manager
(fixed in a preceding commit).

Reviewed-on: http://gerrit.cloudera.org:8080/6637
Tested-by: Kudu Jenkins

+++ src/kudu/clock/hybrid_clock.cc
@@ -251,7 +251,8 @@ bool HybridClock::HasPhysicalComponent() const {
 }
 
 MonoDelta HybridClock::GetPhysicalComponentDifference(Timestamp lhs, Timestamp rhs) const {
-  return MonoDelta::FromMicroseconds(GetPhysicalValueMicros(lhs) - GetPhysicalValueMicros(rhs));
+  return MonoDelta::FromMicroseconds(static_cast<int64_t>(GetPhysicalValueMicros(lhs)) -
+                                     static_cast<int64_t>(GetPhysicalValueMicros(rhs)));
 }
 
 Status HybridClock::WaitUntilAfter(const Timestamp& then,
@@ -459,8 +460,9 @@ Timestamp HybridClock::TimestampFromMicrosecondsAndLogicalValue(
 
 Timestamp HybridClock::AddPhysicalTimeToTimestamp(const Timestamp& original,
                                                   const MonoDelta& to_add) {
-  uint64_t new_physical = GetPhysicalValueMicros(original) + to_add.ToMicroseconds();
-  uint64_t old_logical = GetLogicalValue(original);
+  int64_t new_physical = static_cast<int64_t>(GetPhysicalValueMicros(original))
+                       + to_add.ToMicroseconds();
+  int64_t old_logical = GetLogicalValue(original);
   return TimestampFromMicrosecondsAndLogicalValue(new_physical, old_logical);
 }
